Revert "Added validators for transaction and transaction receipt in PendingState"
This reverts commit 2b1db0f0b6f8c6935782fb0890b8f860f29e90e7.
