+++ ethereumj-core/src/main/java/org/ethereum/casper/config/CasperProperties.java
@@ -1,7 +1,13 @@
 package org.ethereum.casper.config;
 
+import org.apache.commons.lang3.tuple.Pair;
+import org.ethereum.casper.config.net.CasperTestNetConfig;
+import org.ethereum.config.BlockchainNetConfig;
 import org.ethereum.config.SystemProperties;
+import org.ethereum.core.genesis.GenesisConfig;
 import org.ethereum.util.ByteUtil;
+import org.ethereum.validator.BlockCustomHashRule;
+import org.ethereum.validator.BlockHeaderValidator;
 
 import java.io.BufferedReader;
 import java.io.IOException;
