+++ ethereumj-core/src/main/java/org/ethereum/config/SystemProperties.java
@@ -35,8 +35,8 @@ import org.ethereum.net.rlpx.MessageCodec;
 import org.ethereum.net.rlpx.Node;
 import org.ethereum.util.BuildInfo;
 import org.ethereum.util.ByteUtil;
-import org.ethereum.validator.block.BlockCustomHashRule;
-import org.ethereum.validator.block.BlockHeaderValidator;
+import org.ethereum.validator.BlockCustomHashRule;
+import org.ethereum.validator.BlockHeaderValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spongycastle.util.encoders.Hex;
