+++ ethereumj-core/src/main/java/org/ethereum/config/blockchain/AbstractDaoConfig.java
@@ -17,16 +17,19 @@
  */
 package org.ethereum.config.blockchain;
 
+import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.ethereum.config.BlockchainConfig;
 import org.ethereum.core.BlockHeader;
 import org.ethereum.core.Transaction;
-import org.ethereum.validator.block.BlockHeaderRule;
-import org.ethereum.validator.block.BlockHeaderValidator;
-import org.ethereum.validator.block.ExtraDataPresenceRule;
+import org.ethereum.validator.BlockHeaderRule;
+import org.ethereum.validator.BlockHeaderValidator;
+import org.ethereum.validator.ExtraDataPresenceRule;
 import org.spongycastle.util.encoders.Hex;
 
 import java.math.BigInteger;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * Created by Stan Reshetnyk on 26.12.16.
