+++ ethereumj-core/src/main/java/org/ethereum/core/BlockchainImpl.java
@@ -29,8 +29,8 @@ import org.ethereum.listener.EthereumListener;
 import org.ethereum.listener.EthereumListenerAdapter;
 import org.ethereum.manager.AdminInfo;
 import org.ethereum.util.*;
-import org.ethereum.validator.block.DependentBlockHeaderRule;
-import org.ethereum.validator.block.ParentBlockHeaderValidator;
+import org.ethereum.validator.DependentBlockHeaderRule;
+import org.ethereum.validator.ParentBlockHeaderValidator;
 import org.ethereum.vm.program.invoke.ProgramInvokeFactory;
 import org.ethereum.vm.program.invoke.ProgramInvokeFactoryImpl;
 import org.slf4j.Logger;
