+++ ethereumj-core/src/main/java/org/ethereum/core/PendingStateImpl.java
@@ -40,10 +40,6 @@ import org.ethereum.listener.EthereumListener.PendingTransactionState;
 import org.ethereum.listener.EthereumListenerAdapter;
 import org.ethereum.util.ByteUtil;
 import org.ethereum.util.FastByteComparisons;
-import org.ethereum.validator.EntityValidator;
-import org.ethereum.validator.transaction.TransactionReceiptValidator;
-import org.ethereum.validator.transaction.TransactionRule;
-import org.ethereum.validator.transaction.TransactionValidator;
 import org.ethereum.vm.program.invoke.ProgramInvokeFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -83,12 +79,6 @@ public class PendingStateImpl implements PendingState {
 
     @Autowired
     private Blockchain blockchain;
-
-    @Autowired
-    private TransactionValidator transactionValidator;
-
-    @Autowired
-    private TransactionReceiptValidator receiptValidator;
 
     private BlockStore blockStore;
 
@@ -207,8 +197,7 @@ public class PendingStateImpl implements PendingState {
     }
 
     protected boolean receiptIsValid(TransactionReceipt receipt) {
-        EntityValidator.ValidationResult res = receiptValidator.validate(receipt);
-        return res.success;
+        return receipt.isValid();
     }
 
     public synchronized void trackTransaction(Transaction tx) {
@@ -248,13 +237,13 @@ public class PendingStateImpl implements PendingState {
         TransactionReceipt newReceipt = new TransactionReceipt();
         newReceipt.setTransaction(tx);
 
-        TransactionRule.ValidationResult res = transactionValidator.validate(tx);
+        String err = validate(tx);
 
         TransactionReceipt txReceipt;
-        if (res.success) {
-            txReceipt = executeTx(tx);
+        if (err != null) {
+            txReceipt = createDroppedReceipt(tx, err);
         } else {
-            txReceipt = createDroppedReceipt(tx, res.error);
+            txReceipt = executeTx(tx);
         }
 
         if (!receiptIsValid(txReceipt)) {
@@ -481,13 +470,5 @@ public class PendingStateImpl implements PendingState {
 
     public void setCommonConfig(CommonConfig commonConfig) {
         this.commonConfig = commonConfig;
-    }
-
-    public void setTransactionValidator(TransactionValidator transactionValidator) {
-        this.transactionValidator = transactionValidator;
-    }
-
-    public void setReceiptValidator(TransactionReceiptValidator receiptValidator) {
-        this.receiptValidator = receiptValidator;
     }
 }
