+++ ethereumj-core/src/main/java/org/ethereum/net/eth/handler/Eth62.java
@@ -35,8 +35,8 @@ import org.ethereum.sync.SyncManager;
 import org.ethereum.sync.PeerState;
 import org.ethereum.sync.SyncStatistics;
 import org.ethereum.util.ByteUtil;
-import org.ethereum.validator.block.BlockHeaderRule;
-import org.ethereum.validator.block.BlockHeaderValidator;
+import org.ethereum.validator.BlockHeaderRule;
+import org.ethereum.validator.BlockHeaderValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spongycastle.util.encoders.Hex;
