+++ ethereumj-core/src/main/java/org/ethereum/sync/BlockDownloader.java
@@ -22,7 +22,8 @@ import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import org.ethereum.core.*;
 import org.ethereum.net.server.Channel;
-import org.ethereum.validator.block.BlockHeaderValidator;
+import org.ethereum.util.ByteArrayMap;
+import org.ethereum.validator.BlockHeaderValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spongycastle.util.encoders.Hex;
