+++ ethereumj-core/src/main/java/org/ethereum/sync/SyncManager.java
@@ -26,7 +26,7 @@ import org.ethereum.listener.EthereumListener;
 import org.ethereum.net.server.Channel;
 import org.ethereum.net.server.ChannelManager;
 import org.ethereum.util.ExecutorPipeline;
-import org.ethereum.validator.block.BlockHeaderValidator;
+import org.ethereum.validator.BlockHeaderValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spongycastle.util.encoders.Hex;
