+++ ethereumj-core/src/main/java/org/ethereum/validator/BlockCustomHashRule.java
@@ -15,27 +15,29 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.validator.block;
+package org.ethereum.validator;
 
 import org.ethereum.core.BlockHeader;
-import org.ethereum.validator.block.BlockHeaderRule;
-
-import java.math.BigInteger;
+import org.ethereum.util.FastByteComparisons;
+import org.spongycastle.util.encoders.Hex;
 
 /**
- * Checks {@link BlockHeader#gasUsed} against {@link BlockHeader#gasLimit}
- *
- * @author Mikhail Kalinin
- * @since 02.09.2015
+ * Created by Stan Reshetnyk on 26.12.16.
  */
-public class GasValueRule extends BlockHeaderRule {
+public class BlockCustomHashRule extends BlockHeaderRule {
+
+    public final byte[] blockHash;
+
+    public BlockCustomHashRule(byte[] blockHash) {
+        this.blockHash = blockHash;
+    }
 
     @Override
     public ValidationResult validate(BlockHeader header) {
-        if (new BigInteger(1, header.getGasLimit()).compareTo(BigInteger.valueOf(header.getGasUsed())) < 0) {
-            return fault("header.getGasLimit() < header.getGasUsed()");
+        if (!FastByteComparisons.equal(header.getHash(), blockHash)) {
+            return fault("Block " + header.getNumber() + " hash constraint violated. Expected:" +
+                    Hex.toHexString(blockHash) + ", got: " + Hex.toHexString(header.getHash()));
         }
-
         return Success;
     }
 }
