+++ ethereumj-core/src/main/java/org/ethereum/validator/BlockHashRule.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.validator.block;
+package org.ethereum.validator;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.ethereum.config.BlockchainNetConfig;
