+++ ethereumj-core/src/main/java/org/ethereum/validator/BlockHeaderRule.java
@@ -15,36 +15,18 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.validator.block;
+package org.ethereum.validator;
 
 import org.ethereum.core.BlockHeader;
-import org.ethereum.validator.AbstractValidationRule;
 import org.slf4j.Logger;
-
-import java.util.LinkedList;
-import java.util.List;
 
 /**
  * Parent class for {@link BlockHeader} validators
- * which run depends on the header of another block
  *
  * @author Mikhail Kalinin
  * @since 02.09.2015
  */
-public abstract class DependentBlockHeaderRule extends AbstractValidationRule {
-
-    protected List<String> errors = new LinkedList<>();
-
-    public List<String> getErrors() {
-        return errors;
-    }
-
-    public void logErrors(Logger logger) {
-        if (logger.isErrorEnabled())
-            for (String msg : errors) {
-                logger.warn("{} invalid: {}", getEntityClass().getSimpleName(), msg);
-            }
-    }
+public abstract class BlockHeaderRule extends AbstractValidationRule {
 
     @Override
     public Class getEntityClass() {
@@ -54,10 +36,36 @@ public abstract class DependentBlockHeaderRule extends AbstractValidationRule {
     /**
      * Runs header validation and returns its result
      *
-     * @param header block's header
-     * @param dependency header of the dependency block
-     * @return true if validation passed, false otherwise
+     * @param header block header
      */
-    abstract public boolean validate(BlockHeader header, BlockHeader dependency);
+    abstract public ValidationResult validate(BlockHeader header);
 
+    protected ValidationResult fault(String error) {
+        return new ValidationResult(false, error);
+    }
+
+    public static final ValidationResult Success = new ValidationResult(true, null);
+
+    public boolean validateAndLog(BlockHeader header, Logger logger) {
+        ValidationResult result = validate(header);
+        if (!result.success && logger.isErrorEnabled()) {
+            logger.warn("{} invalid {}", getEntityClass(), result.error);
+        }
+        return result.success;
+    }
+
+    /**
+     * Validation result is either success or fault
+     */
+    public static final class ValidationResult {
+
+        public final boolean success;
+
+        public final String error;
+
+        public ValidationResult(boolean success, String error) {
+            this.success = success;
+            this.error = error;
+        }
+    }
 }
