+++ ethereumj-core/src/main/java/org/ethereum/validator/GasValueRule.java
@@ -15,29 +15,26 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.validator.transaction;
+package org.ethereum.validator;
 
-import org.ethereum.core.TransactionReceipt;
-import org.ethereum.validator.EntityValidator;
+import org.ethereum.core.BlockHeader;
 
-import java.util.List;
+import java.math.BigInteger;
 
 /**
- * Composite {@link TransactionReceipt} validator
- * aggregating list of simple validation rules
+ * Checks {@link BlockHeader#gasUsed} against {@link BlockHeader#gasLimit}
+ *
+ * @author Mikhail Kalinin
+ * @since 02.09.2015
  */
-public class TransactionReceiptValidator extends EntityValidator<TransactionReceiptRule, TransactionReceipt> {
-
-    public TransactionReceiptValidator(List<TransactionReceiptRule> rules) {
-        super(rules);
-    }
-
-    public TransactionReceiptValidator(TransactionReceiptRule... rules) {
-        super(rules);
-    }
+public class GasValueRule extends BlockHeaderRule {
 
     @Override
-    public Class getEntityClass() {
-        return TransactionReceipt.class;
+    public ValidationResult validate(BlockHeader header) {
+        if (new BigInteger(1, header.getGasLimit()).compareTo(BigInteger.valueOf(header.getGasUsed())) < 0) {
+            return fault("header.getGasLimit() < header.getGasUsed()");
+        }
+
+        return Success;
     }
 }
