+++ ethereumj-core/src/main/java/org/ethereum/validator/ParentNumberRule.java
@@ -15,40 +15,25 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.validator.block;
+package org.ethereum.validator;
 
-import org.ethereum.config.Constants;
-import org.ethereum.config.SystemProperties;
 import org.ethereum.core.BlockHeader;
 
 /**
- * Checks diff between number of some block and number of our best block. <br>
- * The diff must be more than -1 * {@link Constants#getBEST_NUMBER_DIFF_LIMIT}
+ * Checks if {@link BlockHeader#number} == {@link BlockHeader#number} + 1 of parent's block
  *
  * @author Mikhail Kalinin
  * @since 02.09.2015
  */
-public class BestNumberRule extends DependentBlockHeaderRule {
-
-    private final int BEST_NUMBER_DIFF_LIMIT;
-
-    public BestNumberRule(SystemProperties config) {
-        BEST_NUMBER_DIFF_LIMIT = config.getBlockchainConfig().
-                getCommonConstants().getBEST_NUMBER_DIFF_LIMIT();
-    }
+public class ParentNumberRule extends DependentBlockHeaderRule {
 
     @Override
-    public boolean validate(BlockHeader header, BlockHeader bestHeader) {
+    public boolean validate(BlockHeader header, BlockHeader parent) {
 
         errors.clear();
 
-        long diff = header.getNumber() - bestHeader.getNumber();
-
-        if (diff > -1 * BEST_NUMBER_DIFF_LIMIT) {
-            errors.add(String.format(
-                    "#%d: (header.getNumber() - bestHeader.getNumber()) <= BEST_NUMBER_DIFF_LIMIT",
-                    header.getNumber()
-            ));
+        if (header.getNumber() != (parent.getNumber() + 1)) {
+            errors.add(String.format("#%d: block number is not parentBlock number + 1", header.getNumber()));
             return false;
         }
