--- ethereumj-core/src/main/java/org/ethereum/validator/block/ExtraDataRule.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) [2016] [ <ether.camp> ]
- * This file is part of the ethereumJ library.
- *
- * The ethereumJ library is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * The ethereumJ library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
- */
-package org.ethereum.validator.block;
-
-import org.ethereum.config.Constants;
-import org.ethereum.config.SystemProperties;
-import org.ethereum.core.BlockHeader;
-import org.ethereum.validator.block.BlockHeaderRule;
-
-/**
- * Checks {@link BlockHeader#extraData} size against {@link Constants#getMAXIMUM_EXTRA_DATA_SIZE}
- *
- * @author Mikhail Kalinin
- * @since 02.09.2015
- */
-public class ExtraDataRule extends BlockHeaderRule {
-
-    private final int MAXIMUM_EXTRA_DATA_SIZE;
-
-    public ExtraDataRule(SystemProperties config) {
-        MAXIMUM_EXTRA_DATA_SIZE = config.getBlockchainConfig().
-                getCommonConstants().getMAXIMUM_EXTRA_DATA_SIZE();
-    }
-
-    @Override
-    public ValidationResult validate(BlockHeader header) {
-        if (header.getExtraData() != null && header.getExtraData().length > MAXIMUM_EXTRA_DATA_SIZE) {
-            return fault(String.format(
-                    "#%d: header.getExtraData().length > MAXIMUM_EXTRA_DATA_SIZE",
-                    header.getNumber()
-            ));
-        }
-
-        return Success;
-    }
-}
