--- ethereumj-core/src/main/java/org/ethereum/validator/block/ProofOfWorkRule.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) [2016] [ <ether.camp> ]
- * This file is part of the ethereumJ library.
- *
- * The ethereumJ library is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * The ethereumJ library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
- */
-package org.ethereum.validator.block;
-
-import org.ethereum.core.BlockHeader;
-import org.ethereum.util.FastByteComparisons;
-import org.ethereum.validator.block.BlockHeaderRule;
-
-/**
- * Checks proof value against its boundary for the block header
- *
- * @author Mikhail Kalinin
- * @since 02.09.2015
- */
-public class ProofOfWorkRule extends BlockHeaderRule {
-
-    @Override
-    public ValidationResult validate(BlockHeader header) {
-        byte[] proof = header.calcPowValue();
-        byte[] boundary = header.getPowBoundary();
-
-        if (!header.isGenesis() && FastByteComparisons.compareTo(proof, 0, 32, boundary, 0, 32) > 0) {
-            return fault(String.format("#%d: proofValue > header.getPowBoundary()", header.getNumber()));
-        }
-
-        return Success;
-    }
-}
